requirements.txt
Django==4.2.6
python-decouple==3.8
djangorestframework~=3.14.0
